Major revisions including fixing hanging with large files
- converts PDF to PNGs at lower resolution (72); this resolves issues with barcode recognition (higher resolutions results in missed barcodes

- multiple options for barcode acquisition: adds code for relying on `zbarimg` (shell executable via `barcodeScan_zbarimg`) as an alternative to `barcodeScan_python_zbar_sub` -- note that using python's zbar library seems to be a bit faster (306 vs. 326 sec in a quick-and-dirty comparison)

- multiple options for handling image: adds code for relying on `cv2` for handling early image manipulation as an alternative to relying on PIL (306 vs. 302 sec for cv2 and pil, respectively, suggests that there isn't much of a difference between the two)

- renames json_barcode_not_found_msg as json_msg_barcode_not_found

- removes residual busca-related code: busca_using_png_gen, busca_deskew_scanset, busca_handle_scoring, busca_pages_in_test, and calc_start_pages

- more on hanging: 

No problems w/invocation from python. However, invocation from lisp resulted in early termination of pdfxcb. Example: `(tt::burst-and-log "/home/thomp/Google.Drive.thompfpu/academic/courses/physiol--human/scans/2018/final/scans-c--1000920.pdf" 688 :log-level 0 :output-dir (tt::scans-dir 688))`. Hung at page 149 of 160. No indication of completion in log. The `pdfxcb` process is still running but STAT (`ps -u --pid ...`) for both `zbarimg` and `pdfxcb` shows Sl (process asleep).

+++ pdfxcb/pdfxcb.py
@@ -97,10 +97,10 @@ def locate_cover_sheets (png_files,containing_dir,match_re):
         lg.info(json1.json_progress("looking for barcode on page " + str(i) + " of " + str(i_max) + " png files"))
         image_file_spec = os.path.join(containing_dir,png_files[i])
         lg.debug(image_file_spec)
-        scan_region = ([0,0,0.5,0.5])
+        scan_region = ([0,0,0.7,0.5])
         maybe_barcode = barScan.barcodeScan(
             image_file_spec,
-            scan_region
+            scan_region         # None
         )
         # don't ignore barcode if consider is true
         consider = True
@@ -170,9 +170,10 @@ def pdfxcb (pdf_file_spec,output_dir,match_re):
     cover_sheet_barcodes, cover_sheet_indices = locate_cover_sheets(png_files,output_dir,match_re)
     lg.debug(cover_sheet_barcodes)
     lg.debug(cover_sheet_indices)
-    # clean up PNG files
-    for png_file in png_files:
-        os.remove(os.path.join(output_dir,png_file))
+    clean_up_png_files = False # True
+    if clean_up_png_files:
+        for png_file in png_files:
+            os.remove(os.path.join(output_dir,png_file))
     # write PDFs
     pdf_length = len(png_files)
     page_ranges = generate_page_ranges(cover_sheet_indices,pdf_length)
@@ -189,177 +190,6 @@ def pdfxcb (pdf_file_spec,output_dir,match_re):
              }
     ))
     return True
-
-
-# if IGNORE_PAGE_P is True, skip the scanned page following the pages of each test (for double-sided scanning with a test with an odd number of pages)
-
-# OLD: PDF_FILE_SPEC is either a list of strings (specifying multiple pdf files) or a string specifying a single file.
-
-# NEW: PDF_FILE_SPEC is a string specifying a single PDF file (any 'concatenation' of multiple PDF files must be handled elsewhere...)
-def busca_using_png_gen(pdf_file_spec, xy_file, ignore_page_p):
-    """Deskew and perform OMR on bubbles in file or files specified by PDF_FILE_SPEC, using XY_FILE as guide to bubble locations. Write OMR results, as files of type 'bsc', to the directory holding the file specified by PDF_FILE_SPEC. Note that the file specified by PDF_FILE_SPEC may hold scans for multiple responses/tests from multiple users."""
-    global lg
-    lg.debug("BUSCA.00 (busca_using_png_gen)")
-    # DEBUGP should be either True or False (any other value will interpreted as False)
-    debugp = False
-    blank_page_triggers_error_p = True
-    sanity_checks(ignore_page_p,xy_file)
-    pages_in_test = busca_pages_in_test(xy_file)
-    scanset_pngs_generator = pdf.pdf_to_pngs_generator(pdf_file_spec,pages_in_test,ignore_page_p)
-    scanSets=[]
-    loop_p = True
-    while loop_p:
-        try:
-            # numbering of pages begins at 1
-            png_files,first_page_number,last_page_number = scanset_pngs_generator.next()
-            lg.debug("BUSCA.60 %s %s %s",png_files,first_page_number,last_page_number)
-            scanset = {
-                'bscPath': None,
-                'pdfPageStart': first_page_number,
-                'pdfPageEnd': last_page_number,
-                'pdfPath': pdf_file_spec,
-                'pngPaths': png_files,
-                'deskewPaths': [],
-                'deskewStatus': None,
-                'gradeitStatus': None,
-                'instid': None,
-                'score': None,
-            }
-            scanSets.append( scanset )
-            # handle deskewing
-            busca_deskew_scanset(scanset,blank_page_triggers_error_p,debugp)
-            busca_handle_scoring(scanset,xy_file)
-            # offer consumer of log data the opportunity to deal with scansets as they become available
-            lg.info(json1.json_scanset(scanset))
-        except StopIteration:
-            lg.debug("BUSCA.A0")
-            loop_p = False
-    scanSets_json = json1.json_scansets(scanSets)
-    lg.info(scanSets_json)
-    return True
-
-def busca_deskew_scanset (scanSet,blank_page_triggers_error_p,debugp):
-    page_numbers = [n for n in range(scanSet['pdfPageStart'],scanSet['pdfPageEnd']+1)]
-    png_files = scanSet['pngPaths']
-    for page_number, png_file in zip(page_numbers,png_files):
-        lg.debug("png_file: %s", png_file)
-        # if png_file is None, we treat it as an indication to ignore that position in png_files
-        if png_file:
-            try:
-                # deskew_img_file returns None on an apparent blank page
-                deskew_file = deskew.deskew_img_file(png_file, debugp)
-                if (blank_page_triggers_error_p and deskew_file == None):
-                    scanSet['deskewStatus'] = { 'code': 121, 'message': "encountered blank page on attempt to deskew", 'file': png_file }
-                    lg.info(json1.json_blank_page_on_deskew(png_file))
-                else:
-                    scanSet['deskewStatus'] = { 'code': 20, 'message': "successful deskew", 'file': png_file }
-                    lg.info(json1.json_successful_deskew(png_file))
-            except AssertionError as e:
-                lg.debug("BUSCA_DESKEW_SCANSET.A0: %s %s %s %s",e,e.__doc__,e.message,e.args)
-                raise e
-            except exceptions.CalibrationMarkNotDetected as e:
-                lg.debug("BUSCA_DESKEW_SCANSET.A10: %s %s %s", e, e.__doc__, e.args)
-                lg.debug("BUSCA_DESKEW_SCANSET.A11: %s", e.message)
-                lg.debug("BUSCA_DESKEW_SCANSET.A12: %s", e.diagnostic_file)
-                lg.debug("BUSCA_DESKEW_SCANSET.A13: %s", e.additional_notes)
-                comments = [ e.message ]
-                if e.additional_notes:
-                    comments.extend(e.additional_notes)
-                lg.error(json1.json_failed_to_deskew(e.diagnostic_file, page_number, comments))
-                scanSet['deskewStatus'] = { 'code': 120, 'message': "failed to deskew", 'file': png_file, 'page': page_number }
-                # as long as we log the above, it's desirable to tuck this aside and process everything else
-                # sys.exit("failed to deskew file"+file);
-                # FIXME: do we want this? or do we simply want to not append None below?
-                deskew_file = None
-            except Exception as e:
-                lg.debug("BUSCA_DESKEW_SCANSET.A2: %s %s %s %s", e, e.__doc__, e.message, e.args)
-                lg.error(json1.json_failed_to_deskew(png_file, page_number, e.message))
-                scanSet['deskewStatus'] = { 'code': 120, 'message': "failed to deskew", 'file': png_file, 'page': page_number }
-                # as long as we log the above, it's desirable to tuck this aside and process everything else
-                # sys.exit("failed to deskew file"+file);
-                # FIXME: do we want this? or do we simply want to not append None below?
-                deskew_file = None
-        else:
-            deskew_file = None
-            lg.debug("deskew_file: %s", deskew_file)
-        scanSet['deskewPaths'].append(deskew_file)
-    lg.debug("BUSCA.B0")
-    if (not (scanSet['deskewPaths'] and len(scanSet['deskewPaths'])>0)):
-        lg.error(json1.json_failed_to_deskew('', '', 'Failed to generate deskew file(s) from PNG files -- this should not happen -- all deskew failures should be handled on a per-file basis...'))
-        lg.info(json1.json_last_log_msg())
-        sys.exit("Aborted busca -- see the system log")
-
-def busca_handle_scoring (scanSet,xy_file):
-    deskew_file_set = scanSet['deskewPaths']
-    lg.debug("DESKEW_FILE_SET: %s", deskew_file_set)
-    # anticipate possibility that deskew failed for one or more files in the set
-    if None in deskew_file_set:
-        pass
-    # if None isn't present, assume deskewing was successful
-    else:
-        lg.debug("BUSCA.D0")
-        page_scores = []
-        # this is where we want to catch an issue with a single scanset -- e.g., if bubbles can't be located
-        try:
-            barcodes = bubbles.barcodes_on_pages(deskew_file_set)
-            lg.debug("BUSCA.D1: barcodes: %s",barcodes)
-            # FIXME: 2015-09-24 -- something in bubbles.score_pages(...) was dying w/o error
-            page_scores,bubble_detection_failed_p,diagnosticFiles = bubbles.score_pages(deskew_file_set, xy_file)
-            lg.debug("BUSCA.D2: %s %s",bubble_detection_failed_p,diagnosticFiles)
-            if bubble_detection_failed_p:
-                messageString="Failed to locate anticipated bubbles -- see diagnostic file(s) at"
-                for diagnosticFile in diagnosticFiles:
-                    messageString=messageString+" "+diagnosticFile
-                scanSet['gradeitStatus'] = { 'code': 61, 'message':  messageString, 'files': diagnosticFiles, 'page': ''}
-            lg.debug("BUSCA.D3")
-        except bubbles.LocateBubbleError as e:
-            lg.debug("BUSCA LocateBubbleError: %s",e)
-            # note: this is generated by bubbles.score_bubbles --> maybe better to have it caught at calling function so page can be specified?
-            messageString="Failed to locate anticipated bubbles -- see diagnostic file(s) at"
-            for efile in e.files:
-                messageString = messageString+" "+efile
-            scanSet['gradeitStatus'] = { 'code': 61, 'message':  messageString, 'files': e.files, 'page': ''}
-        except:
-            # !! if this occurs, something needs to be debugged -> don't ignore it; ensure this shows up in the log irrespective of other considerations
-            msg = 'Unexpected error: {}'.format(sys.exc_info()[0])
-            traceback_string = traceback.format_exc()
-            lg.critical(msg)
-            lg.critical(traceback_string)
-            lg.critical(json1.json_msg(255, msg, False,None))
-        lg.debug("BUSCA.E0")
-        input_file_sans_suffix, input_file_suffix = os.path.splitext(scanSet['pdfPath'])
-        # FIXME: this only scales well up to 999 responses
-        index_format_string = "{0:0>03d}"
-        index_string = str.format(index_format_string,scanSet['pdfPageStart'])
-        if page_scores:
-            bsc_file_suffix = 'bsc'
-            bsc_file = input_file_sans_suffix + index_string + "." + bsc_file_suffix
-            write_page_scores(page_scores, bsc_file)
-            scanSet['bscPath'] = bsc_file
-        scanSet['barcodes'] = barcodes
-        paths_file = input_file_sans_suffix + index_string + ".paths"
-        write_paths(deskew_file_set, paths_file)
-
-def busca_pages_in_test (xy_file):
-    try:
-        # FIXME: rename as pages_per_set ?
-        pages_in_test = bubbles.xy_to_page_number(xy_file)
-    except Exception as e:
-        lg.error(json1.json_failed_to_parse_file(e,xy_file))
-        lg.info(json1.json_last_log_msg())
-        sys.exit("No point in proceeding if we can't parse the XY file to determine the number of pages... " + xy_file)
-    if pages_in_test < 1:
-        lg.error(json1.json_zero_page_test(xy_file))
-        lg.info(json1.json_last_log_msg())
-        sys.exit("Doesn't make much sense to scan a test that is composed of zero pages...")
-    return pages_in_test
-
-
-
-def calc_start_pages (total_scanned_pages, pages_in_set):
-    """Return a list of integers representing page numbers where tests begin (page numbering begins with 0)."""
-    start_pages = [ i for i in range(0, total_scanned_pages) if i % pages_in_set == 0 ]
-    return start_pages
 
 def directory_sanity_check (directory_spec,exitp):
     if not os.path.isdir(directory_spec):
@@ -402,6 +232,7 @@ def module_sanity_check (module_name,exitp):
 
 def split_pdf_to_png_files (pdf_file_spec,output_dir):
     """Split PDF file specified by PDF_FILE_SPEC into a series of files, each representing a single page as a PNG image. Write files to directory specified by OUTPUT_DIR."""
+    png_files = None
     try:
         # sanity check
         assert os.path.isabs(pdf_file_spec)
@@ -508,8 +339,15 @@ def main():
         match_re = re.compile(match_re_string)
     pdf_file_spec = args.input_files[0]
     lg.debug(pdf_file_spec)
-    pdfxcb(pdf_file_spec,args.output_dir,match_re)
-    # busca_using_png_gen(pdf_files[0], xy_file, ignore_page_p)
+    # generic debugging
+    lg.debug(os.getcwd())         # current/working directory
+    # might also want to import platform to get architecture, other details...
+    try:
+        pdfxcb(pdf_file_spec,args.output_dir,match_re)
+    except Exception as e:
+        lg.error("Crash and burn")
+        lg.error(sys.exc_info()[0])
+        raise
     lg.info(json1.json_last_log_msg())
 
 if __name__ == "__main__":
